+++ test/FacadeAct.test.ts
@@ -258,7 +258,7 @@ describe('FacadeAct contract', () => {
 
     it('Basket - Should handle no valid basket after refresh', async () => {
       // Redeem all RTokens
-      await rToken.connect(addr1).redeem(issueAmount)
+      await rToken.connect(addr1).redeem(issueAmount, true)
 
       // Set simple basket with only one collateral
       await basketHandler.connect(owner).setPrimeBasket([aToken.address], [fp('1')])
@@ -528,7 +528,7 @@ describe('FacadeAct contract', () => {
 
     it('Revenues - Should handle assets with invalid claim logic', async () => {
       // Redeem all RTokens
-      await rToken.connect(addr1).redeem(issueAmount)
+      await rToken.connect(addr1).redeem(issueAmount, true)
 
       // Setup a new aToken with invalid claim data
       const ATokenCollateralFactory = await ethers.getContractFactory(